+++ src/starbase/__init__.py
@@ -1,8 +1,8 @@
-__title__ = 'starbase.__init__'
-__version__ = '0.3.1'
-__build__ = 0x00000b
+__title__ = 'starbase'
+__version__ = '0.3.2'
+__build__ = 0x00000c
 __author__ = 'Artur Barseghyan'
-__copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
+__copyright__ = 'Copyright (c) 2013-2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('Connection', 'Table',)