+++ src/starbase/client/transport/__init__.py
@@ -1,18 +1,28 @@
 __title__ = 'starbase.client.transport.__init__'
 __author__ = 'Artur Barseghyan'
-__copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
+__copyright__ = 'Copyright (c) 2013-2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('HttpRequest', 'HttpResponse')
 
 import json
+import logging
+import time
+
 import requests
 from requests.auth import HTTPBasicAuth
+from requests.exceptions import RequestException
 
 from six import string_types
 
 from starbase.json_decoder import json_decode
 from starbase.content_types import MEDIA_TYPE_JSON
-from starbase.client.transport.methods import GET, PUT, POST, DELETE, METHODS, DEFAULT_METHOD
+from starbase.client.transport.methods import (
+    GET, PUT, POST, DELETE, METHODS, DEFAULT_METHOD
+)
+from starbase.client.transport import status_codes
+from starbase.translations import _
+
+logger = logging.getLogger(__name__)
 
 class HttpResponse(object):
     """
@@ -40,14 +50,17 @@ class HttpResponse(object):
         except:
             return self.raw.code
 
-    def get_content(self, decode_content=False, keys_to_bypass_decoding=[], keys_to_skip=[]):
+    def get_content(self, decode_content=False, keys_to_bypass_decoding=[], \
+                    keys_to_skip=[]):
         """
         Gets response content.
 
-        :param bool decode_content: If set to True, content is decoded with default decoder, having
-            the empty keys ignored.
-        :param list|tuple|set keys_to_bypass_decoding: List of keys to bypass decoding.
-        :param list|tuple|set keys_to_skip: List of keys to ignore (won't be in the resulted content).
+        :param bool decode_content: If set to True, content is decoded with
+            default decoder, having the empty keys ignored.
+        :param list|tuple|set keys_to_bypass_decoding: List of keys to bypass
+            decoding.
+        :param list|tuple|set keys_to_skip: List of keys to ignore (won't be
+            in the resulted content).
         :return str:
         """
         if self.code == 200: # TODO - is it?
@@ -72,7 +85,8 @@ class HttpRequest(object):
     :param str method:
     :param bool fail_silently:
     """
-    def __init__(self, connection, url='', data={}, decode_content=False, method=DEFAULT_METHOD, fail_silently=True):
+    def __init__(self, connection, url='', data={}, decode_content=False, \
+                 method=DEFAULT_METHOD, fail_silently=True):
         """
         See the docs above.
         """
@@ -85,7 +99,8 @@ class HttpRequest(object):
         self.fail_silently = fail_silently
         headers = {
             'Accept': str(self.__connection.content_type),
-            'Content-type': str(self.__connection.content_type) + '; charset=UTF-8'
+            'Content-type': str(self.__connection.content_type) + \
+                            '; charset=UTF-8'
             }
         endpoint_url = self.__connection.base_url + self.url
 
@@ -107,17 +122,44 @@ class HttpRequest(object):
             request_data['data'] = data
 
         if self.__connection.user and self.__connection.password:
-            request_data['auth'] = HTTPBasicAuth(self.__connection.user, self.__connection.password)
+            request_data['auth'] = HTTPBasicAuth(
+                self.__connection.user,
+                self.__connection.password
+                )
 
-        # For the sake of simplicity the `requests` library replaced the `urllib2`.
+        for i in range(connection.retries + 1):
+            self.response = self.call(method, request_data)
+
+            if self.response.status_code in status_codes.BAD_STATUS_CODES:
+                # If status code is 4** or 5** or raise an exception
+                # means, retry.
+                if i < connection.retries:
+                    delay = connection.retry_delay * (2**i)
+                    logger.warn(
+                        _("Hbase returned error: {0}. Sleeping "
+                          "for {1} seconds").format(self.response.reason, delay)
+                        )
+                    time.sleep(delay)
+                else:
+                    return None
+            else:
+                return None
+
+    def call(self, method, request_data):
+        """
+        For the sake of simplicity the `requests` library replaced the
+        `urllib2`.
+
+        Moved to seperate function to aid mocking in tests
+        """
         if GET == method:
-            self.response = requests.get(**request_data)
+            return requests.get(**request_data)
         elif PUT == method:
-            self.response = requests.put(**request_data)
+            return requests.put(**request_data)
         elif POST == method:
-            self.response = requests.post(**request_data)
+            return requests.post(**request_data)
         elif DELETE == method:
-            self.response = requests.delete(**request_data)
+            return requests.delete(**request_data)
 
     def get_response(self):
         """
@@ -138,9 +180,10 @@ class HttpRequest(object):
             except Exception as e:
                 pass
         else:
-            raise NotImplementedError("Connection type {0} is not implemented.".format(
-                self.__connection.content_type
-                ))
+            raise NotImplementedError(
+                "Connection type {0} is "
+                "not implemented.".format(self.__connection.content_type)
+                )
 
         if self.decode_content and self.response.ok: # Make sure OK is ok.
             response_content = json_decode(response_content)