+++ src/starbase/exceptions.py
@@ -1,6 +1,6 @@
 __title__ = 'starbase.exceptions'
 __author__ = 'Artur Barseghyan'
-__copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
+__copyright__ = 'Copyright (c) 2013-2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'BaseException', 'ImproperlyConfigured', 'InvalidArguments', 'ParseError', 'DoesNotExist',