+++ test.sh
@@ -3,4 +3,5 @@ reset
 clear
 ./install.sh
 clear
+python src/starbase/json_decoder/tests.py
 python src/starbase/client/tests.py